% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FixationFilterFunctions.R
\name{downsample_gaze}
\alias{downsample_gaze}
\title{Downsample gaze}
\usage{
downsample_gaze(data_in, ds.factor, xcol = "x", ycol = "y")
}
\arguments{
\item{data_in}{Data frame which must contain the variables specified by the parameters xcol and ycol.}

\item{ds.factor}{The factor to down-sample by. For example, setting ds.factor to 10 down-samples data recorded at 1000 HZ to 100 HZ.}

\item{xcol}{Name of the column where raw x values are stored. Default: x}

\item{ycol}{Name of the column where raw y values are stored. Default: y}
}
\value{
Data frame with downsampled gaze data. The output variables are x, y, and the numbers of the first and last samples of the original data frame included in the bin.
}
\description{
This function downs-samples gaze by a specified factor. Data are down-sampled by splitting the data in bins and calculating the mean of each bin.
}
