% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FixationFilterFunctions.R
\name{find.transition.weights}
\alias{find.transition.weights}
\title{Find transition weights for each sample in a gaze matrix.}
\usage{
find.transition.weights(data_in, window.step.size = 6, windowsize)
}
\arguments{
\item{data_in}{Input data}

\item{window.step.size}{Step size}

\item{windowsize}{Window size}
}
\value{
transition weights.
}
\description{
This function is used internally by the function algorithm_i2mc
}
