% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreprocessingFunctions.R
\name{interpolate_with_margin}
\alias{interpolate_with_margin}
\title{Interpolate over gaps (subsequent NAs) in vector.}
\usage{
interpolate_with_margin(data_in, marg, max_gap)
}
\arguments{
\item{data_in}{Vector to interpolate in}

\item{marg}{Margin in samples before and after gap to use for interpolation}

\item{max_gap}{Maximum length of gaps in sample}
}
\value{
vector with interpolated gaps
}
\description{
Interpolate over gaps (subsequent NAs) in vector.
}
