% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{sample.data.classified}
\alias{sample.data.classified}
\title{Sample-to-sample raw and fixation classified data from 1 individual}
\format{
A data frame
\describe{
  \item{x}{fixation position x}
  \item{y}{fixation position y}
  \item{x.raw}{fixation position x}
  \item{y.raw}{fixation position y}
  \item{timestamp}{timestamp in milliseconds}
  }
}
\source{
The dataset was stored in the package at 'data/example_data.RData'
}
\usage{
sample.data.classified
}
\description{
This dataset contains sample-to-sample data from 1 individuals during a free viewing tasks. Data were
recorded at 1200 Hz using a Tobii Pro Spectrum eye tracker. Fixations were classified with the I-VT algorithm with a velocity threshold set to 30 degrees/seconds
and default settings in the function algorithm_ivt
}
\keyword{datasets}
