% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{sample.data.filtered}
\alias{sample.data.filtered}
\title{Fixation-filtered sample-by-sample example data}
\format{
A data frame
\describe{
  \item{timestamp}{timestamp in ms recorded by the eye tracker}
  \item{x.raw}{unfiltered gaze position x}
  \item{y.raw}{unfiltered gaze position y}
  \item{x}{fixation filtered gaze position x}
  \item{y}{fixation filtered gaze position y}
}
}
\source{
The dataset was stored in the package at 'data/example_data.RData'
}
\usage{
sample.data.filtered
}
\description{
This dataset contains data from 1 individuals during a free viewing tasks after pre-processing. Data were
recorded at 1200 Hz using a Tobii Pro Spectrum eye tracker
}
\keyword{datasets}
