% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{sample.data.processed}
\alias{sample.data.processed}
\title{Pre-processed sample-by-sample example data}
\format{
A data frame
\describe{
  \item{id}{participant number}
  \item{timestamp}{timestamp in ms recorded by the eye tracker}
  \item{x.raw}{gaze position x}
  \item{y.raw}{gaze position y}
  \item{x.unprocessed}{copy of gaze position x before preprocessing}
  \item{y.unprocessed}{copy of gaze position y before preprocessing}
  \item{timestamp.t}{"'Theoretical timestamp' for comparison."}
  \item{sample}{sample nr in recording}
}
}
\source{
The dataset was stored in the package at 'data/example_data.RData'
}
\usage{
sample.data.processed
}
\description{
This dataset contains data from 1 individuals during a free viewing tasks after pre-processing. Data were
recorded at 1200 Hz using a Tobii Pro Spectrum eye tracker
}
\keyword{datasets}
