% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{sample.data.saccades}
\alias{sample.data.saccades}
\title{Saccades from 3 individuals}
\format{
A data frame
\describe{
  \item{onset}{onset of the saccade in ms}
  \item{x.onset}{gaze position x at onset}
  \item{y.onset}{gaze position y at onset }
  \item{offset}{offset of the saccade in ms}
  \item{x.offset}{gaze position x at offset}
  \item{y.offset}{gaze position y at offset }
  \item{duration}{duration of saccade in ms}
  \item{amplitude}{amplitude of saccade in degrees}
  \item{peak.velocity}{peak velocity of saccade}
  \item{velocity.profile}{velocity profile}
  \item{id}{participant number}
  }
}
\source{
The dataset was stored in the package at 'data/example_data.RData'
}
\usage{
sample.data.saccades
}
\description{
This dataset contains saccade data from 3 individuals during a free viewing tasks. Data were
recorded at 1200 Hz using a Tobii Pro Spectrum eye tracker
}
\keyword{datasets}
