% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{sample.data.unprocessed}
\alias{sample.data.unprocessed}
\title{Unprocessed sample-by-sample example data}
\format{
A data frame
\describe{
  \item{id}{participant number}
  \item{timestamp}{timestamp in ms recorded by the eye tracker}
  \item{x.raw}{gaze position x}
  \item{y.raw}{gaze position y}
}
}
\source{
The dataset was stored in the package at 'data/example_data.RData'
}
\usage{
sample.data.unprocessed
}
\description{
This dataset contains data from 3 individuals during a free viewing tasks before pre-processing. Data were
recorded at 1200 Hz using a Tobii Pro Spectrum eye tracker
}
\keyword{datasets}
