% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FixationFilterFunctions.R
\name{summarize_fixation_metrics}
\alias{summarize_fixation_metrics}
\title{Summarize fixation statistics}
\usage{
summarize_fixation_metrics(
  fixation.candidate.starts,
  fixation.candidate.stops,
  x,
  y,
  timestamp,
  one_degree = 40
)
}
\arguments{
\item{fixation.candidate.starts}{First row in the data included in the fixation}

\item{fixation.candidate.stops}{Last row in the data included in the fixation}

\item{x}{X coordinates}

\item{y}{Y coordinates}

\item{timestamp}{Timestamps in milliseconds}

\item{one_degree}{one degree of the visual field in the unit of the x and y coordinates in the data. Typically pixels or degrees.}
}
\value{
data frame with fixation descriptives
}
\description{
Summarize descriptives for a fixation defined by onset and offset rows in the data. Used internally by event classification functions.
}
